grammar to understand input
The grammar files are in antlr/Sequence.g and
antlr/SequenceWalker.g. They can be compiled from the same
directory with the following command:

	./antlr-compile *.g

Currently, the output files are soft-linked into the
appropriate src directory.

This grammar is only for training. The input must be a
sequence of words and terminators, where words are
lowercase ascii characters, and terminators are one of
the sequences defined in Sequence.g.