initial "strict" mode commit
Trainer and Agent are wrappers for CharacterMap.

A CharacterMap can be trained with a sequence of inputs and
expected outputs with Trainer.suggest(). This needs to
match the grammar for a "sequence" in Sequence.g.

The CharacterMap will then try to map a stream of inputs
back to the expected output with Agent.feed(). This
method returns any definitely-matched strings up to that
point.