agent now uses basic statistics
Core functionality is implemented in Agent.java. To train
it, call Agent.train() with an ordered list of inputs and
their corresponding expected output string. To allow non-
exact matches, call Agent.addUncertainty() with a positive
integer. Using a lower value will cause it to more highly
favor matches seen in testing.

The agent can then be tested (or used) by calling
Agent.test() when a single input. This method returns an
array as large as its memory (given when an Agent is
instantiated). The first element of the buffer is the
expected output for that single character, the next element
is the expected output for the previous character, and so
on.